Deleted spreadsheet_utils after removing column index functions
