+++ webtools/hedweb/constants/file_constants.py
@@ -7,3 +7,4 @@ TEXT_FILE_EXTENSIONS = ['.txt', '.tsv']
 SCHEMA_WIKI_EXTENSION = '.mediawiki'
 SCHEMA_XML_EXTENSION = '.xml'
 SCHEMA_EXTENSIONS = [SCHEMA_XML_EXTENSION, SCHEMA_WIKI_EXTENSION]
+FILE_EXTENSION_TO_DELIMITER_DICTIONARY = {'.txt': '\t', '.tsv': '\t'}
