--- webtools/hedweb/constants/spreadsheet_constants.py
@@ -1,10 +1,0 @@
-OTHER_TAG_COLUMN_NAMES = ['Event Details', 'Multiple Tags', 'HED tag', 'HED tags', 'Tag', 'Tags']
-SPECIFIC_TAG_COLUMN_NAMES = ['Category', 'Description', 'Label', 'Long']
-SPECIFIC_TAG_COLUMN_NAMES_DICTIONARY = {'Category': ['Categories', 'Category', 'Event Categories', 'Event Category'],
-                                        'Description': ['Description', 'Descriptions', 'Description in text',
-                                                        'Event Description', 'Event Descriptions'],
-                                        'Label': ['Event Label', 'Event Labels', 'Label', 'Labels', 'Short',
-                                                  'Short Label', 'Short Labels'],
-                                        'Long': ['Long', 'Long name', 'Long names']}
-SPREADSHEET_FILE_EXTENSION_TO_DELIMITER_DICTIONARY = {'.txt': '\t', '.tsv': '\t'}
-
